flask/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.8.2
cachetools==5.4.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
dnspython==2.3.0
Flask==3.0.3
fonttools==4.41.0
frozenlist==1.4.0
geographiclib==2.0
geopy==2.4.1
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-vision==3.7.3
googleapis-common-protos==1.63.2
googlemaps==4.10.0
grpcio==1.65.1
grpcio-status==1.65.1
idna==3.4
importlib-resources==6.0.0
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.3.1
kiwisolver==1.4.4
logic==0.2.3
MarkupSafe==2.1.5
matplotlib==3.7.2
multidict==6.0.4
multipledispatch==1.0.0
numpy==1.25.1
openai==0.27.8
opencv-python-headless==4.10.0.84
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
plotly==5.15.0
proto-plus==1.24.0
protobuf==5.27.2
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pymongo==4.4.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
sklearn==0.0.post5
tenacity==8.2.2
threadpoolctl==3.2.0
toolz==0.12.1
tqdm==4.65.0
tzdata==2023.3
unification==0.2.2
urllib3==2.0.3
Werkzeug==3.0.3
yarl==1.9.2
zipp==3.16.1
